One more float sequence fix

+++ Run.py
@@ -122,7 +122,7 @@ def main():
             # Bring in next job into every position in sequence_list
             init_jobs += 1
             init_job_list = list(range(init_jobs))
-            new_sequence_list = np.array([]).reshape(0, init_jobs)
+            new_sequence_list = np.array([], dtype=int).reshape(0, init_jobs)
             for s in sequence_list:
                 for pos in range(s.size + 1):
                     new_sequence_list = np.vstack((new_sequence_list, np.insert(s, pos, init_jobs - 1)))
